mm_project/requirements.txt
Django~=5.1.1
python-dateutil~=2.9.0.post0
djangorestframework~=3.15.2
django-crispy-forms
psycopg2-binary
whitenoise
django-permissions-policy
crispy-bootstrap5